Refactor form syntax (#15)
* Rename Form to Page

The name "Form" could confuse users with the whole form, which is the
set of pages.

* Move button text param and always return dict

Button text param is better suited in the run method because it is
the last called method as the 'next' button is the last component to be
displayed.

Form will always return a dict instead of returning the single value when
there is only one input. This behavior was made in order to keep the
compatible behavior with read functions outside Form.
Now each read function extracts the single value from answer dict.